Add search tracking
Use the Design System approach for tracking search interactions. When
the search text input is interacted with or a search result is clicked
on, an event will be added to the `dataLayer`. The code is based on the
following from alphagov#govuk-design-system:

- https://github.com/alphagov/govuk-design-system/blob/main/src/javascripts/components/analytics.mjs
- https://github.com/alphagov/govuk-design-system/blob/main/src/javascripts/components/search.tracking.mjs

with some changes because of how the search implementation differs in
x-govuk#govuk-eleventy-plugin (web component).